+++ .gitignore
@@ -2,3 +2,4 @@ target
 dependency-reduced-pom.xml
 .idea
 *.iml
+.env
